add proxies to all handlersn
