+++ src/notte_eval/agent_handlers/__init__.py
@@ -1,5 +1,22 @@
 import importlib
+from enum import StrEnum
 from typing import Any, NamedTuple
+
+from pydantic import BaseModel
+
+
+class PoolEnum(StrEnum):
+    NONE = "None"
+    ANCHOR = "Anchor"
+    STEEL = "Steel"
+    BROWSERBASE = "BrowserBase"
+    CAMOUFOX = "Camoufox"
+
+
+class Proxy(BaseModel):
+    server: str
+    username: str
+    password: str
 
 
 def fetch_handler(key: str) -> tuple[type, type]:
